+++ src/App.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from 'react'
 import { Container, Row, Col } from 'react-bootstrap'
-import Countdown from './components/Countdown'
-import Footer from './components/Footer'
-import Header from './components/Header'
-import LPTable from './components/LPSection'
-import MintSection from './components/MintSection'
-import PriceCard from './components/PriceCard'
-import StockTable from './components/StockSection'
+import Countdown from './components/display/Countdown'
+import Footer from './components/layout/Footer'
+import Header from './components/layout/Header'
+import LPTable from './components/display/lp/LPSection'
+import MintSection from './components/display/mint/MintSection'
+import PriceCard from './components/display/PriceCard'
+import StockTable from './components/display/stock/StockSection'
 import { getDopplePrice, getTwinPrice } from './modules/GovernanceToken'
 
 const App = () => {
