+++ src/components/layout/Header.tsx
@@ -1,6 +1,6 @@
 import { useState, useCallback } from 'react'
 import { Container, Nav, Navbar, Form, Button, InputGroup } from 'react-bootstrap'
-import { getAddressInQueryString } from '../modules/Utils'
+import { getAddressInQueryString } from '../../modules/Utils'
 
 const AddressForm = () => {
   const [address, setAddress] = useState(getAddressInQueryString() ?? '')
